% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sqlf_config_edit}
\alias{sqlf_config_edit}
\title{Open the sqlfluff configuration file for editing}
\usage{
sqlf_config_edit(path = ".sqlfluff")
}
\arguments{
\item{path}{Path to the configuration file. Defaults to `".sqlfluff"`.}
}
\value{
The file path, invisibly.
}
\description{
Opens the `.sqlfluff` configuration file in the default editor.
In RStudio, this opens the file in the source pane.
}
\examples{
\dontrun{ # interactive option to edit config file
sqlf_config_edit()
}

}
