% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sqlf_dialects}
\alias{sqlf_dialects}
\title{List available SQL dialects}
\usage{
sqlf_dialects()
}
\value{
A data.frame with columns `label`, `name`, and `inherits_from`.
}
\description{
Returns the dialects supported by the installed version of sqlfluff.
}
\examples{
\dontshow{if (reticulate::virtualenv_exists("r-sqlfluffr")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sqlf_dialects()
\dontshow{\}) # examplesIf}
}
