% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sqlf_reset}
\alias{sqlf_reset}
\title{Reset sqlfluffr setup}
\usage{
sqlf_reset()
}
\value{
Invisible `NULL`.
}
\description{
Removes the cached marker file so that the next call to any sqlfluffr
function will re-prompt for Python and sqlfluff installation. Use this
if your Python environment has changed or something is not working.
}
\examples{
\dontrun{ # Resets cached install marker
sqlf_reset()
}

}
