% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sqlf_rules}
\alias{sqlf_rules}
\title{List available linting rules}
\usage{
sqlf_rules()
}
\value{
A data.frame with columns `code` and `description`.
}
\description{
Returns the rules available in the installed version of sqlfluff.
}
\examples{
\dontshow{if (reticulate::py_module_available("sqlfluff")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
sqlf_rules()
\dontshow{\}) # examplesIf}
}
