% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sqlf_uninstall}
\alias{sqlf_uninstall}
\title{Uninstall sqlfluff}
\usage{
sqlf_uninstall(confirm = interactive())
}
\arguments{
\item{confirm}{If `TRUE` (default in interactive sessions), ask for
confirmation before removing.}
}
\value{
Invisible `NULL`.
}
\description{
Removes the virtual environment created by [sqlf_install()].
}
