% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlfluffr-package.R
\docType{package}
\name{sqlfluffr-package}
\alias{sqlfluffr}
\alias{sqlfluffr-package}
\title{sqlfluffr: Wrapper to the SQL Linter and Formatter 'sqlfluff'}
\description{
Provides an R interface to the 'Python' 'sqlfluff' SQL linter and formatter via the 'reticulate' package. Lint, fix, and parse SQL queries with support for multiple dialects. 'Python' and 'sqlfluff' are automatically provisioned on first use with user consent. Includes special handling for 'glue' SQL syntax with curly-brace placeholders.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brendensm/sqlfluffr}
  \item Report bugs at \url{https://github.com/brendensm/sqlfluffr/issues}
}

}
\author{
\strong{Maintainer}: Brenden Smith \email{brendensmithmi@gmail.com} [copyright holder]

}
\keyword{internal}
